feat(index) use new render function
