+++ src/index.html
@@ -9,7 +9,7 @@
   <meta name="description" content="Yet another chat">
 </head>
 
-<body>
-  <script type="module" src="./index2.ts"></script>
+<body id="root">
+  <script type="module" src="./index.ts"></script>
 </body>
 </html>
