--- src/index2.ts
@@ -1,6 +1,0 @@
-import './index.css';
-import Login from './pages/login';
-
-const loginPage = new Login({});
-
-document.body.appendChild(loginPage.getContent());
