+++ src/lib/renderDOM.ts
@@ -1,0 +1,6 @@
+export default function render(query, block) {
+  const root = document.querySelector(query);
+
+  root.appendChild(block.getContent());
+  return root;
+}
